+++ docs/changelog.rst
@@ -27,6 +27,7 @@ Fixes
 * Plotter with colorbar work with recipes (no longer recursion in ``to_dict()``)
 * :class:`aspecd.processing.SliceRemoval` handles lists properly
 * :class:`aspecd.processing.SliceRemoval` removes value(s) from corresponding axis
+* :class:`aspecd.processing.Averaging` handles inverted axes (*e.g.*, ppm scale) correctly regardless how ranges are given
 
 
 Version 0.11.0